# Improving performance of generic map copy