--- a/pkg/config/generic_map_test.go
+++ b/pkg/config/generic_map_test.go
@@ -1,0 +1,17 @@
+package config
+
+import (
+	"fmt"
+	"testing"
+)
+
+func BenchmarkGenericMap_Copy(b *testing.B) {
+	m := GenericMap{}
+	for i := 0; i < 20; i++ {
+		m[fmt.Sprintf("key-%d", i)] = fmt.Sprintf("value-%d", i)
+	}
+
+	for i := 0; i < b.N; i++ {
+		_ = m.Copy()
+	}
+}