The GenericMap's Copy method is run in diverse stages of the pipeline, for each processed flow.

Preallocating the map size to the expected resulting size can improve Copy CPU performance and memory allocations by 40%-50%.

Benchmark results with 20-entries maps (before vs after):

```
BenchmarkGenericMap_Copy-12       658083              1671 ns/op            1852 B/op          2 allocs/op
BenchmarkGenericMap_Copy-12       978686              1082 ns/op            1199 B/op          1 allocs/op
```

Benchmark results with 30-entries maps (before vs after):

```
BenchmarkGenericMap_Copy-12       354102              3280 ns/op            4337 B/op          4 allocs/op
BenchmarkGenericMap_Copy-12       633978              1758 ns/op            2324 B/op          1 allocs/op
```