@mariomac 2022-10-14
@jpinsonneau if you know the approximate length of the resulting map, yes (same for slices).

If you don't know the length, doing `m := map[string]string{}` is commonly preferred.